talk-to-me/pyproject.toml
[project]
name = "talk_to_me"
version = "0.1.0"
description = "A2A-powered multi-agent conversation system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[all]",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-ollama>=0.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "client", "config"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]